Mass renaming for consistency
Previously, I used "runtime" to describe both the base implementation and
the (not yet existing) facilities exposed to basic lisp code. Failing to
recognize a distinction between these two will make lisp package naming in
particular difficult to do in a way that makes sense. Now, "virtual machine"
will be used to describe the C and lisp code that runs exclusively on existing
systems (i.e. the UEFI environment and SBCL) and "runtime" will be used to
describe the virtual machine intrinsics available to applications along with the
lisp code that abstracts over the different execution environments.

"Virtual machine" itself previously referred to the bytecode interpreter. This
is now just the "bytecode interpreter."