Fix a bug in the handling of immediates in object files
The object file code previously checked if a reference was a pointer and assumed
it to be an immediate otherwise. This logic is incorrect as some encodings
correspond to neither a pointer nor an immediate. Specifically, the value 7 was
being treated by BORAX_DISCRIMINATE as equivalent to
BORAX_DISCRIM_OBJECT_RECORD, which led to an invalid memory read.

It is tempting to have BORAX_DISCRIMINATE look for encodings that are not valid
immediates and return an ILLEGAL discriminator, but this would complicate the
garbage collector solely to handle an invariant violation. Instead, object files
that violate the memory system's invariants should be rejected.